test: rename directory for serde test cases
